Fix for issues with login and pull behavior added in version 3.0. (#5)
Fix for a couple of issue with the login and pull behavior added in version 3.0. First, there was an issue with the conditional in the "pull" step. Then, the role was performing the registry login
AFTER the pull step, meaning that if you wanted to pull in an image from a registry that you are
logging into, it wouldn't work. Reordering the operations solved the problem.